+++ search/search-client-elasticsearch/src/main/java/io/camunda/search/es/transformers/ElasticsearchTransformer.java
@@ -12,10 +12,10 @@ import co.elastic.clients.elasticsearch._types.SortOptions;
 import co.elastic.clients.elasticsearch._types.aggregations.Aggregate;
 import co.elastic.clients.elasticsearch._types.aggregations.Aggregation;
 import co.elastic.clients.elasticsearch._types.query_dsl.Query;
-import io.camunda.search.clients.aggregation.SearchAggregate;
-import io.camunda.search.clients.aggregation.SearchAggregation;
 import co.elastic.clients.elasticsearch.core.search.SourceConfig;
 import co.elastic.clients.elasticsearch.core.search.SourceFilter;
+import io.camunda.search.clients.aggregation.SearchAggregate;
+import io.camunda.search.clients.aggregation.SearchAggregation;
 import io.camunda.search.clients.query.SearchQuery;
 import io.camunda.search.clients.source.SearchSourceConfig;
 import io.camunda.search.clients.source.SearchSourceFilter;
