+++ search/search-client-elasticsearch/src/main/java/io/camunda/search/es/transformers/ElasticsearchTransformers.java
@@ -11,9 +11,10 @@ import co.elastic.clients.elasticsearch._types.aggregations.Aggregate;
 import co.elastic.clients.elasticsearch._types.aggregations.CardinalityAggregate;
 import co.elastic.clients.elasticsearch._types.aggregations.LongTermsAggregate;
 import io.camunda.search.clients.aggregation.SearchAggregation;
+import io.camunda.search.clients.aggregation.SearchCardinalityAggregate;
 import io.camunda.search.clients.aggregation.SearchCardinalityAggregation;
-import io.camunda.search.clients.aggregator.SearchTermsAggregator;
 import io.camunda.search.clients.aggregation.SearchTermsAggregation;
+import io.camunda.search.clients.aggregator.SearchTermsAggregator;
 import io.camunda.search.clients.core.SearchQueryHit;
 import io.camunda.search.clients.core.SearchQueryRequest;
 import io.camunda.search.clients.core.SearchQueryResponse;
@@ -38,11 +39,10 @@ import io.camunda.search.clients.transformers.SearchTransfomer;
 import io.camunda.search.clients.types.TypedValue;
 import io.camunda.search.es.transformers.aggregation.CardinalityAggregateTransformer;
 import io.camunda.search.es.transformers.aggregation.CardinalityAggregationTransformer;
-import io.camunda.search.es.transformers.aggregator.TermsAggregationTransformer;
 import io.camunda.search.es.transformers.aggregation.SearchAggregateTransformer;
 import io.camunda.search.es.transformers.aggregation.SearchAggregationTransformer;
 import io.camunda.search.es.transformers.aggregation.TermsAggregateTransformer;
-import io.camunda.search.es.transformers.aggregation.TermsAggregationTransformer;
+import io.camunda.search.es.transformers.aggregator.TermsAggregationTransformer;
 import io.camunda.search.es.transformers.query.BoolQueryTransformer;
 import io.camunda.search.es.transformers.query.ConstantScoreQueryTransformer;
 import io.camunda.search.es.transformers.query.ExistsQueryTransformer;
