+++ search/search-client-elasticsearch/src/main/java/io/camunda/search/es/transformers/aggregation/AggregateOptionTransformer.java
@@ -9,9 +9,9 @@ package io.camunda.search.es.transformers.aggregation;
 
 import co.elastic.clients.elasticsearch._types.aggregations.AggregateVariant;
 import io.camunda.search.clients.aggregation.SearchAggregateOption;
+import io.camunda.search.clients.transformers.SearchTransfomer;
 import io.camunda.search.es.transformers.ElasticsearchTransformer;
 import io.camunda.search.es.transformers.ElasticsearchTransformers;
-import io.camunda.search.transformers.SearchTransfomer;
 
 public abstract class AggregateOptionTransformer<
         T extends AggregateVariant, R extends SearchAggregateOption>
