+++ search/search-client-elasticsearch/src/main/java/io/camunda/search/es/transformers/aggregation/SearchAggregateTransformer.java
@@ -11,9 +11,9 @@ import co.elastic.clients.elasticsearch._types.aggregations.Aggregate;
 import co.elastic.clients.elasticsearch._types.aggregations.AggregateVariant;
 import io.camunda.search.clients.aggregation.SearchAggregate;
 import io.camunda.search.clients.aggregation.SearchAggregateOption;
+import io.camunda.search.clients.transformers.SearchTransfomer;
 import io.camunda.search.es.transformers.ElasticsearchTransformer;
 import io.camunda.search.es.transformers.ElasticsearchTransformers;
-import io.camunda.search.transformers.SearchTransfomer;
 
 public class SearchAggregateTransformer
     extends ElasticsearchTransformer<Aggregate, SearchAggregate> {
