+++ search/search-client-query-transformer/src/main/java/io/camunda/search/clients/aggregation/SearchCardinalityAggregation.java
